[EUnit] Implement model comparison for model references and HUTN models
While working on the last bits of the ETL example for the MODELS
article, I noticed that HutnModel and ModelReferences had issues when
performing comparisons:

- ModelReference could be used on the left-hand side of a comparison,
  as it implemented IComparableModel. However, ModelReference couldn't
  be on the right-hand side of a comparsion, as AbstractEmfModel didn't
  know how to follow it.

- HutnModel didn't implement model comparison at all.

I implemented IComparableModel in HutnModel by delegating to its inner
EMF model. That ensured comparisons worked if a HutnModel was on the left-hand
side.

For the cases when ModelReferences or HutnModels were on the right-hand side,
I needed a different approach. I didn't want to have AbstractEmfModel or
assert*EqualModels() follow the ModelReferences or unwrap the HutnModels
by themselves.

For this reason, I have defined a new subinterface of IModel called
IAdaptableModel, with a single method: T adaptTo(Class<T>). It returns a non-null
value if the model (a ModelReference or an HutnModel) can be adapted to the target
type (AbstractEmfModel). ModelReference implements it by trying to convert the
target model to the target type, and HutnModel tries to convert the inner EMF model
to the target type.

This ensures regular EMF models, HutnModels and ModelReferences can be freely
mixed when doing comparisons with assert*EqualModels().

This commit also adds new test cases for model comparisons between ModelReferences,
HutnModels and regular EMF models.