[EUnit] Do not fall back on StringBasedDiffViewer if delta != null
This is useful to disable the Compare button when differences between
models have been detected, but they come from having an empty expected
or actual model. The EMF Compare difference viewer does not handle
empty models well, and converting the models to strings will not help
either. We cannot use file comparison either, as that'd require to
save the models to temporary files, which is wasteful: we already know
which model is empty, after all.