ExeedEditor: do not use '*' mapping for IDXMIResourceFactoryImpl
Using this mapping breaks Exeed for the Papyrus UML metamodels: Activities will
not show any children in it. However, this works fine with the Sample Reflective
Ecore Editor, which Exeed is based on.

This is because we override the default '*' mapping to the Eclipse registry, and
neither XMIResourceFactoryImpl nor XMLResourceFactoryImpl seem to be good defaults.
Our best option is to drop the '*' mapping and move it to 'model', which is the
"generic" file extension we propose in most of our documentation.