emc.emf: avoid naming conflicts in isOfKind and isOfType (fix bug 382666)
There are two kinds of type references in EOL: absolute type references (using
one or more segments separated with :: for every EPackage and the final EClass)
and relative type references (just the EClass name).

Absolute references take into account the EPackages that the EClass is in, but
relative references only contain the EClass, so there might be naming conflicts
if we simply look through all available EPackages as we usually did. For instance,
both the UML and the xText metamodels have an "Action" EClass.

One way to do it would be to look up and store all available types in the resource,
but it would be too expensive. As a middle-ground approach, we now try first with
the EClass objects in the EPackage of the first EObject in the EMF resource, if it
exists. That should be good enough to solve most naming conflicts.

If that doesn't work, we will follow the old approach of trying all available EPackages.