InMemoryEmfModel: new EPackage registry if explicit list of EPackages is provided (fix bug 382666)
InMemoryEmfModel sets up its resource set's EPackage registry as follows:

- When no EPackages are mentioned in its constructor, it uses the global
  EPackage registry.

- When some EPackages are mentioned, it adds them to the existing EPackage
  registry in the resource set of the wrapped resource.

Only adding the EPackages to the existing registry might not work as expected,
as in some situations that may be the global EPackage registry. For instance,
Exeed's InMemoryEmfModel objects are created without mentioning any EPackages,
and EWL's InMemoryEmfModel may end up using that global registry as well,
producing the name conflicts mentioned in bug #382666.

This commit modifies InMemoryEmfModel so whenever an explicit list of EPackages
are passed through the constructor, it sets up a new EPackageRegistryImpl that
delegates to the global EPackage registry. I tried having it create the new
EPackageRegistryImpl all the time, but that broke quite a few tests.