Split off eunit.dt.cmp.emf and eunit.dt.diff.emf into a separate feature
Kepler uses EMF Compare 2.x instead of EMF Compare 1.x. EMF Compare 2.x
is not backwards compatible with EMF Compare 1.x, so this breaks all the
code we have using it.

For the time being, we've decided to split off the EMF Compare 1.x bits
into a separate feature, and create an alternative version that uses
EMF Compare 2.x.