[EUnit] Add support for comparator options in 'assertEqualModels' and fix EuGENia regression tests
I needed to ignore differences due to whitespace in the EuGENia regression
tests, so I extended assertEqualModels with another argument that takes a
Map {...} and configures the comparator with it. It can be used like this:

    assertEqualModels('A', 'B', Map { 'whitespace' = 'ignore' })

or

    assertEqualModels('My message', 'A', 'B', Map { 'whitspace' = 'ignore' })

With this and several 1-line changes in the .gmfgen files, all tests pass
again.