[Eugenia] Fix for bug #307755
I noticed this as well for other delegates. The main problem was that models
are disposed of in the same order as they are loaded. Since the ECore model
was disposed of *before* the GenModel model, the GenModel model was in an
inconsistent state and saved itself as an empty resource.