Multiple uses of metamodel{File,URI} should always result in the same EPackage instances
I noticed that one of the reasons the model comparisons with EMF Compare 2.x
were failing was that it relied on models of the same metamodel using the
exact same EPackage instance. This would be naturally honored for a permanently
registered metamodel, but it didn't work for the metamodels temporarily registered
through the metamodelFile and metamodelURI attributes of the epsilon.emf.loadModel
task.

I found out that this was due to two things:

- EmfModel was registering the metamodel on the resource set's package registry,
  which may not be shared with later EmfModels. It should be registering the
  metamodel in the global EMF registry, which is shared with all EmfModels.

- EmfUtil#register was always replacing the instance in the registry, instead of
  first checking if something was already there. It should reuse the existing
  instance if it were already registered (e.g. from a previous epsilon.emf.loadModel
  task).

This commit fixes these two issues, so the Eugenia tests work now, except for the
missing difference filtering options that were not ported to the new version of
the comparator.