Report a change during full build.
This assumes that everything is changed and needs a build.
All Concrodance listeners (e.g. database indexer) are checked and have proper implementations available.

Note that using delete/add might result in problems when listeners react on the delete and destroy their own resources (e.g. close an editor), which was not intented by the full build.